Refactoring: function to create RegionalTaskRefactoring: moving excess code from R5RCore to other classes - Focus on DetailedItineraryPlanner
